+++ src/pages/PaymentsSuccess.tsx
@@ -9,6 +9,7 @@ import Button from 'wowds-ui/Button';
 import GlobalSize from '@/constants/globalSize';
 import RoutePath from '@/routes/routePath';
 import { color } from 'wowds-tokens';
+import ordersApi from '@/apis/orders/ordersApi';
 
 export function PaymentsSuccess() {
   const navigate = useNavigate();
@@ -16,51 +17,34 @@ export function PaymentsSuccess() {
   const [responseData, setResponseData] = useState(null);
 
   useEffect(() => {
-    const requestData = {
-      orderId: searchParams.get('orderId'),
-      amount: searchParams.get('amount'),
-      paymentKey: searchParams.get('paymentKey')
-    };
-
-    // TODO: 개발자센터에 로그인해서 내 결제위젯 연동 키 > 시크릿 키를 입력하세요. 시크릿 키는 외부에 공개되면 안돼요.
-    // @docs https://docs.tosspayments.com/reference/using-api/api-keys
-    const secretKey = 'test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6';
-
-    // 토스페이먼츠 API는 시크릿 키를 사용자 ID로 사용하고, 비밀번호는 사용하지 않습니다.
-    // 비밀번호가 없다는 것을 알리기 위해 시크릿 키 뒤에 콜론을 추가합니다.
-    // @docs https://docs.tosspayments.com/reference/using-api/authorization#%EC%9D%B8%EC%A6%9D
-    const encryptedSecretKey = `Basic ${btoa(secretKey + ':')}`;
-
     async function confirm() {
-      const response = await fetch(
-        'https://api.tosspayments.com/v1/payments/confirm',
-        {
-          method: 'POST',
-          headers: {
-            Authorization: encryptedSecretKey,
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify(requestData)
-        }
-      );
+      const requestData = {
+        orderId: searchParams.get('orderId'),
+        amount: searchParams.get('amount'),
+        paymentType: searchParams.get('paymentType'),
+        paymentKey: searchParams.get('paymentKey')
+      };
 
-      const json = await response.json();
-
-      if (!response.ok) {
-        // TODO: 구매 실패 비즈니스 로직 구현
-        console.log(json);
-        navigate(`/fail?code=${json.code}&message=${json.message}`);
-        return;
-      }
-
-      // TODO: 결제 성공 비즈니스 로직을 구현하세요.
-      // console.log(json);
+      // Todo: requestData 검증
+      // Todo: 클라이언트에서 들고 있을 데이터 어떻게 관리할지 정하기
+      const json = ordersApi.POST_ORDER({
+        orderNanoId: requestData.orderId || 'hi',
+        membershipId: 1,
+        issuedCouponId: 1,
+        totalAmount: 20000,
+        discountAmount: 0,
+        finalPaymentAmount: requestData.amount ? +requestData.amount : 0
+      });
       return json;
     }
 
-    confirm().then((data) => {
-      setResponseData(data);
-    });
+    confirm()
+      .then((data) => {
+        setResponseData(data);
+      })
+      .catch(() => {
+        navigate(RoutePath.PaymentsFail);
+      });
   }, [searchParams]);
 
   return (
@@ -72,59 +56,6 @@ export function PaymentsSuccess() {
         <Flex direction="column" align="start">
           <Text>이번 학기 회비 결제를 완료했어요.</Text>
           <Text>이제 GDSC 정회원으로 이번 학기에 활동하실 수 있어요!</Text>
-
-          <div className="box_section" style={{ width: '600px' }}>
-            <h2>결제를 완료했어요</h2>
-            <div className="p-grid typography--p" style={{ marginTop: '50px' }}>
-              <div className="p-grid-col text--left">
-                <b>결제금액</b>
-              </div>
-              <div className="p-grid-col text--right" id="amount">
-                {`${Number(searchParams.get('amount')).toLocaleString()}원`}
-              </div>
-            </div>
-            <div className="p-grid typography--p" style={{ marginTop: '10px' }}>
-              <div className="p-grid-col text--left">
-                <b>주문번호</b>
-              </div>
-              <div className="p-grid-col text--right" id="orderId">
-                {`${searchParams.get('orderId')}`}
-              </div>
-            </div>
-            <div className="p-grid typography--p" style={{ marginTop: '10px' }}>
-              <div className="p-grid-col text--left">
-                <b>paymentKey</b>
-              </div>
-              <div
-                className="p-grid-col text--right"
-                id="paymentKey"
-                style={{ whiteSpace: 'initial', width: '250px' }}>
-                {`${searchParams.get('paymentKey')}`}
-              </div>
-            </div>
-            <div className="p-grid-col">
-              <Link to="https://docs.tosspayments.com/guides/payment-widget/integration">
-                <button className="button p-grid-col5">연동 문서</button>
-              </Link>
-              <Link to="https://discord.gg/A4fRFXQhRu">
-                <button
-                  className="button p-grid-col5"
-                  style={{ backgroundColor: '#e8f3ff', color: '#1b64da' }}>
-                  실시간 문의
-                </button>
-              </Link>
-            </div>
-          </div>
-          <div
-            className="box_section"
-            style={{ width: '600px', textAlign: 'left' }}>
-            <b>Response Data :</b>
-            <div id="response" style={{ whiteSpace: 'initial' }}>
-              {responseData && (
-                <pre>{JSON.stringify(responseData, null, 4)}</pre>
-              )}
-            </div>
-          </div>
         </Flex>
       </Flex>
       <Button onClick={() => navigate(RoutePath.MyPage)}>완료하기</Button>
